pyproject.toml
[project]
name = "yolo-booth-detection"
version = "0.1.0"
description = "YOLO-based booth detection project supporting OBB detection and instance segmentation"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "ultralytics>=8.0.0",
    "sahi>=0.11.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "shapely>=2.0.0",
    "pypinyin>=0.49.0",
    "Pillow>=10.0.0",
    "PyMuPDF>=1.23.0",
    "pdf2image>=1.16.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.5.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.eggs
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term"
